Added a SearchController
