+++ src/main/java/com/searchbox/web/HomeController.java
@@ -38,8 +38,7 @@ import com.searchbox.ref.Sort;
 public class HomeController {
 	
 	
-	private static Logger logger = LoggerFactory
-			.getLogger(HomeController.class);
+	private static Logger logger = LoggerFactory.getLogger(HomeController.class);
 
 	@RequestMapping
 	public ModelAndView index(@RequestParam(defaultValue = "unknown") String h) {
@@ -78,54 +77,5 @@ public class HomeController {
 		logger.debug("Here is our param2 " + h);
 		model.addObject("h", h);
 		throw new RuntimeException("wassup");
-	}
-	
-	@RequestMapping("search")
-	public ModelAndView search(HttpServletRequest request) {
-		
-		ModelAndView model = new ModelAndView("search/index");
-
-		model.addObject("title", "Search results");
-		logger.debug("We're in the search controller");
-
-		SearchResult result = new SearchResult();
-		String[] fields = { "title", "description" };
-		result.setFields(Arrays.asList(fields));
-		for (int i = 0; i < 10; i++) {
-			Hit hit = new Hit();
-			hit.getFields().put("title", "here's my " + i + "th Title");
-			String desc = "";
-			for (int t = 0; t < Math.random() * 20; t++) {
-				desc += "lorem Ipsum dolores invictus amenentum centri. ";
-			}
-			hit.getFields().put("description", desc);
-			hit.setScore(new Float(Math.random() + (10 - i)));
-			result.addHit(hit);
-		}
-
-		FieldFacet facet = new FieldFacet("Keyword", "keyword");
-		facet.setOrder(Order.VALUE);
-		facet.setSort(Sort.DESC);
-		facet.addValueElement(facet.new Value("Population", "Population", 29862).setSelected(true));
-		facet.addValueElement("Demographic Factors", 40833);
-		facet.addValueElement("Developing Countries",27923);
-		facet.addValueElement("Research Methodology",25246);
-		facet.addValueElement("Family Planning", 23287);
-		facet.addValueElement("Population Dynamics", 20919);
-
-		FieldFacet facet2 = new FieldFacet("Institution", "institution");
-		facet2.setOrder(Order.KEY);
-		facet2.setSort(Sort.ASC);
-		facet2.addValueElement("Department of Biology, MIT", 647);
-		facet2.addValueElement("Department of Molecular and Cell Biology, University of California, Berkeley.",609);
-		facet2.addValueElement("Division of Biology, California Institute of Technology, Pasadena.",558);
-		facet2.addValueElement("European Molecular Biology Laboratory, Heidelberg, Germany.",543);
-		facet2.addValueElement("ARC", 525);
-		
-		result.addElement(facet);
-		result.addElement(facet2);
-
-		model.addObject("result", result);
-		return model;
 	}
 }
